requirements.txt
transformers>=4.36.2
torch>=2.1.0
sentencepiece>=0.1.99
accelerate>=0.25.0
python-dotenv>=0.19.0
pdfplumber==0.10.3
pytesseract==0.3.10
numpy==1.26.3
pandas==2.1.4
nltk>=3.8.1
python-docx>=0.8.11
PyPDF2>=3.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pytest==7.4.4
black==23.12.1
flake8==7.0.0
langchain>=0.1.0
langchain-core>=0.1.0
streamlit>=1.31.0
Pillow>=10.0.0
python-magic>=0.4.27 